Safe505/requirements.txt
# Web Framework
Flask==3.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Utilidades
cryptography==46.0.5